refactor(db): Phase 1 — structure, transaction_scope, first SQL migrations
Foundation:
- db/tx.py: transaction_scope() as the named entry point for explicit
  transaction boundaries. Alias of get_db_ctx() today, will wrap
  SQLAlchemy Session tomorrow.
- db/repositories/, db/queries/, db/jobs/: empty packages ready for
  the phased migration. New query/write functions go here; existing
  modules stay until touched.

SQL migrations out of routers (baseline 432 → 364, −68 calls):

  api/settings.py (8→0): show counts → db/shows, cache clear → db/cache
  api/tasks.py (8→0): bulk delete → db/tasks
  api/tags.py (3→0): track path lookup → db/library
  api/genres.py (4→0): taxonomy node → db/genres
  api/enrichment.py (9→0): analysis tracks, artist refs, setlist tracks,
    playlist lookup → db/library
  api/radio.py (9→0): album/playlist fetch → db/queries/radio
  api/management.py (10→0): wipe, rebuild, vacuum, storage stats,
    reset analysis → db/management
  api/browse_album.py (9→0): album detail, tracks, similar → db/queries/browse
  api/playlists.py (8→0): playlist create, tracks, reorder → db/playlists

9 router files now have 0 raw SQL. 20 new db functions created.

Tests: fixed 3 radio contract tests broken by the migration (patches
pointed at crate.api.radio.get_db_ctx which no longer exists). 252
tests green.
