+++ app/crate/api/radio.py
@@ -12,10 +12,15 @@ from crate.bliss import (
     generate_virtual_playlist_radio,
 )
 from crate.db import (
-    get_db_ctx,
     get_library_artist_by_id,
     get_library_track_by_storage_id,
     get_user_by_email,
+)
+from crate.db.queries.radio import (
+    get_track_path_by_id,
+    get_track_path_by_pattern,
+    get_album_for_radio,
+    get_playlist_for_radio,
 )
 
 router = APIRouter()
@@ -37,10 +42,7 @@ def _effective_user_id(user: dict) -> int | None:
 
 def _resolve_track_path(track_id: int = 0, path: str = "", storage_id: str = "") -> str | None:
     if track_id:
-        with get_db_ctx() as cur:
-            cur.execute("SELECT path FROM library_tracks WHERE id = %s", (track_id,))
-            row = cur.fetchone()
-        return row["path"] if row else None
+        return get_track_path_by_id(track_id)
 
     if storage_id:
         row = get_library_track_by_storage_id(storage_id)
@@ -49,20 +51,7 @@ def _resolve_track_path(track_id: int = 0, path: str = "", storage_id: str = "")
     if not path:
         return None
 
-    with get_db_ctx() as cur:
-        escaped_path = _escape_like(path)
-        cur.execute(
-            """
-            SELECT path
-            FROM library_tracks
-            WHERE path = %s OR path LIKE %s ESCAPE '\\'
-            ORDER BY CASE WHEN path = %s THEN 0 ELSE 1 END, path ASC
-            LIMIT 1
-            """,
-            (path, f"%{escaped_path}", path),
-        )
-        row = cur.fetchone()
-    return row["path"] if row else None
+    return get_track_path_by_pattern(path, _escape_like(path))
 
 
 _RADIO_CACHE_TTL = 300  # 5 minutes
@@ -149,9 +138,7 @@ def api_track_radio(
 def api_album_radio(request: Request, album_id: int, limit: int = Query(50, ge=1, le=100)):
     user = _require_auth(request)
     effective_user_id = _effective_user_id(user)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT artist, name FROM library_albums WHERE id = %s", (album_id,))
-        row = cur.fetchone()
+    row = get_album_for_radio(album_id)
     if not row:
         raise HTTPException(status_code=404, detail="Album not found")
 
@@ -185,9 +172,7 @@ def api_album_radio(request: Request, album_id: int, limit: int = Query(50, ge=1
 def api_playlist_radio(request: Request, playlist_id: int, limit: int = Query(50, ge=1, le=100)):
     user = _require_auth(request)
     effective_user_id = _effective_user_id(user)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT id, name, scope, user_id, is_active FROM playlists WHERE id = %s", (playlist_id,))
-        row = cur.fetchone()
+    row = get_playlist_for_radio(playlist_id)
     if not row:
         raise HTTPException(status_code=404, detail="Playlist not found")
     if row.get("scope") == "system" and not row.get("is_active", False):
