+++ app/crate/db/playlists.py
@@ -557,3 +557,33 @@ def consume_playlist_invite(token: str) -> dict | None:
         )
         row = cur.fetchone()
     return dict(row) if row else None
+
+
+def get_playlist_filter_options() -> dict:
+    with get_db_ctx() as cur:
+        cur.execute("SELECT DISTINCT format FROM library_tracks WHERE format IS NOT NULL AND format != '' ORDER BY format")
+        formats = [r["format"] for r in cur.fetchall()]
+
+        cur.execute("SELECT DISTINCT audio_key FROM library_tracks WHERE audio_key IS NOT NULL AND audio_key != '' ORDER BY audio_key")
+        keys = [r["audio_key"] for r in cur.fetchall()]
+
+        cur.execute("SELECT DISTINCT audio_scale FROM library_tracks WHERE audio_scale IS NOT NULL AND audio_scale != '' ORDER BY audio_scale")
+        scales = [r["audio_scale"] for r in cur.fetchall()]
+
+        cur.execute("SELECT name FROM library_artists ORDER BY name")
+        artists = [r["name"] for r in cur.fetchall()]
+
+        cur.execute("SELECT MIN(year) AS min_y, MAX(year) AS max_y FROM library_tracks WHERE year IS NOT NULL AND year != ''")
+        yr = cur.fetchone()
+
+        cur.execute("SELECT MIN(bpm) AS min_b, MAX(bpm) AS max_b FROM library_tracks WHERE bpm IS NOT NULL")
+        bpm = cur.fetchone()
+
+    return {
+        "formats": formats,
+        "keys": keys,
+        "scales": scales,
+        "artists": artists,
+        "year_range": [yr["min_y"] or "1960", yr["max_y"] or "2026"],
+        "bpm_range": [int(bpm["min_b"] or 60), int(bpm["max_b"] or 200)],
+    }
