+++ app/tests/test_radio_contracts.py
@@ -76,14 +76,11 @@ class TestRadioApiContracts:
             }
         ]
 
-        with patch("crate.api.radio.get_db_ctx") as mock_ctx, \
-             patch("crate.api.radio.generate_album_radio", return_value=tracks):
-            mock_ctx.return_value.__enter__.return_value.fetchone.return_value = {
+        with patch("crate.api.radio.get_album_for_radio", return_value={
                 "artist": "Converge",
                 "name": "Jane Doe",
-            }
-            mock_ctx.return_value.__exit__.return_value = False
-
+             }), \
+             patch("crate.api.radio.generate_album_radio", return_value=tracks):
             resp = test_app.get("/api/radio/album/5?limit=50")
 
         assert resp.status_code == 200
@@ -105,17 +102,14 @@ class TestRadioApiContracts:
             }
         ]
 
-        with patch("crate.api.radio.get_db_ctx") as mock_ctx, \
-             patch("crate.api.radio.generate_playlist_radio", return_value=tracks):
-            mock_ctx.return_value.__enter__.return_value.fetchone.return_value = {
+        with patch("crate.api.radio.get_playlist_for_radio", return_value={
                 "id": 7,
                 "name": "Hardcore",
                 "scope": "system",
                 "user_id": None,
                 "is_active": True,
-            }
-            mock_ctx.return_value.__exit__.return_value = False
-
+             }), \
+             patch("crate.api.radio.generate_playlist_radio", return_value=tracks):
             resp = test_app.get("/api/radio/playlist/7?limit=50")
 
         assert resp.status_code == 200
@@ -125,16 +119,13 @@ class TestRadioApiContracts:
         assert data["tracks"][0]["title"] == "Jane Doe"
 
     def test_playlist_radio_hides_inactive_system_playlists(self, test_app):
-        with patch("crate.api.radio.get_db_ctx") as mock_ctx:
-            mock_ctx.return_value.__enter__.return_value.fetchone.return_value = {
+        with patch("crate.api.radio.get_playlist_for_radio", return_value={
                 "id": 12,
                 "name": "Hidden Editorial",
                 "scope": "system",
                 "user_id": None,
                 "is_active": False,
-            }
-            mock_ctx.return_value.__exit__.return_value = False
-
+             }):
             resp = test_app.get("/api/radio/playlist/12?limit=50")
 
         assert resp.status_code == 404
